refactor(index): de-junk-drawer the mod.rs spine — extract util/git_context/discovery
index/mod.rs (1158, fan_in 855) is the flagship 'mod.rs is the index, not the
junk drawer' violation. Move three cohesive duties to sibling modules (same
module depth → pub(super)/pub(crate) reach preserved):
  util.rs        now_ms/duration_ms/hex_sha256/path_string(_for_seed)/table_row_count/file_metadata_ms/meta_for
  git_context.rs git_changed_paths/resolve_git_context/live_worktree_contexts/pathspec+path mapping/git_output (+GitChangedPaths)
  discovery.rs   discovery_plan/indexed_file_map/target_for_path (+DiscoveryPlan)
Re-exported pub(crate) (resolve_git_context stays pub for benches). Items called
cross-module bumped pub(crate); IndexFile/DiscoveryPlan fields pub(crate). mod.rs
1158 -> 914. No behavior change.
